Release: Gotham v0.1.0
This is the very first release of Gotham, a static site generator built
on the great Hugo project.

The first release of Gotham is mainly a cosmetic change of Hugo. Sites,
themes, and modules should be 100% compatible between Gotham and Hugo.
No unique features were added in this release.

Gotham v0.1.0 (compatible with Hugo v0.72.0/extended)