Mostly restore original Hugo Magefile
Moving forward, the majority of the changes that we make for Mage to
Gotham can be done in its own file. This allows us to touch Hugo's as
little as possible, making handling conflicts easier in the future.

The parts of the Magefile that continue to be different after this PR
need to be different in order for the Magefile to work for our project.
An additional, Gotham-specific Magefile will come in a separate PR.