Switch to dynamic binaries
A change in Hugo v0.84.x made it so that the static binaries we're
generating are causing a segmentation fault. I haven't been able to
track the specific cause nor do I have a workaround. So for the time
being, Strawberry will revert to dynamic binaries for releases.

Sometime in the future, where we can fix this issue, we'll go back to
static binaries. The GitHub Issue to track the future work can be found
here: https://github.com/strawberryssg/strawberry-v0/issues/278