Release: Strawberry v0.21.0
This release includes the new "Strawberry File" as well as a warning
when building a site not as the root of the hostname. The warning covers
files that only work when they exists at the root. This release bumps
the Hugo base version to 0.90.1.

Strawberry v0.21.0 (compatible with Hugo v0.90.1/extended)